♻️ relocate migration page components from `features/` to `components/`
