+++ frontend/apps/app/components/MigrationDetailPage/components/CopyButton/CopyButton.tsx
