+++ frontend/apps/app/components/MigrationDetailPage/components/CopyButton/index.ts
@@ -1,0 +1,1 @@
+export * from './CopyButton'
