+++ frontend/apps/app/components/MigrationDetailPage/components/MigrationHealthClient/components/RadarChart/index.ts
@@ -1,0 +1,1 @@
+export * from './RadarChart'
