+++ frontend/apps/app/components/MigrationDetailPage/components/MigrationHealthClient/index.ts
@@ -1,0 +1,1 @@
+export * from './MigrationHealthClient'
