+++ frontend/apps/app/components/MigrationDetailPage/components/MigrationHealthClient/utils/calculateScores.ts
@@ -1,6 +1,6 @@
 'use client'
 
-import type { CategoryEnum } from '../components/RadarChart/RadarChart'
+import type { CategoryEnum } from '../components/RadarChart'
 
 type ReviewFeedbackForScore = {
   category: string
