+++ frontend/apps/app/components/MigrationDetailPage/components/ReviewFeedbackList/components/ResolveButton/ResolutionCommentModal/index.ts
@@ -1,0 +1,1 @@
+export * from './ResolutionCommentModal'
