+++ frontend/apps/app/components/MigrationDetailPage/components/ReviewFeedbackList/index.ts
@@ -1,0 +1,1 @@
+export * from './ReviewFeedbackList'
