+++ frontend/apps/app/components/MigrationDetailPage/components/UserFeedbackClient/index.ts
@@ -1,0 +1,1 @@
+export * from './UserFeedbackClient'
