+++ frontend/apps/app/components/MigrationDetailPage/index.ts
